Make FeatureExtractor pluggable
