+++ demos/roms-vss/src/main/resources/application.properties
@@ -2,3 +2,5 @@ spring.mvc.hiddenmethod.filter.enabled=true
 com.redis.om.vss.useLocalImages=false
 com.redis.om.vss.maxLines=300
 redis.om.spring.djl.enabled=true
+server.port=8082
+spring.data.redis.url=redis://172.18.36.1/
