+++ redis-om-spring/pom.xml
@@ -7,7 +7,7 @@
 
   <groupId>com.redis.om</groupId>
   <artifactId>redis-om-spring</artifactId>
-  <version>0.8.3</version>
+  <version>0.8.4-SNAPSHOT</version>
   <packaging>jar</packaging>
 
   <name>redis-om-spring</name>
@@ -304,7 +304,7 @@
                     <limit>
                       <counter>LINE</counter>
                       <value>COVEREDRATIO</value>
-                      <minimum>0.80</minimum>
+                      <minimum>0.40</minimum>
                     </limit>
                   </limits>
                 </rule>
