+++ redis-om-spring/src/main/java/com/redis/om/spring/RedisModulesConfiguration.java
@@ -26,6 +26,7 @@ import com.redis.om.spring.ops.pds.BloomOperations;
 import com.redis.om.spring.search.stream.EntityStream;
 import com.redis.om.spring.search.stream.EntityStreamImpl;
 import com.redis.om.spring.serialization.gson.*;
+import com.redis.om.spring.vectorize.DefaultFeatureExtractor;
 import com.redis.om.spring.vectorize.FeatureExtractor;
 import com.redis.om.spring.vectorize.face.FaceDetectionTranslator;
 import com.redis.om.spring.vectorize.face.FaceFeatureTranslator;
@@ -264,7 +265,7 @@ public class RedisModulesConfiguration {
       @Nullable @Qualifier("redisTemplate") RedisTemplate<?, ?> redisTemplate,
       RedisOMSpringProperties properties,
       ApplicationContext ac) {
-    return properties.getDjl().isEnabled() ? new FeatureExtractor(redisTemplate, ac, imageEmbeddingModel, faceEmbeddingModel, imageFactory, defaultImagePipeline, sentenceTokenizer) : null;
+    return properties.getDjl().isEnabled() ? new DefaultFeatureExtractor(redisTemplate, ac, imageEmbeddingModel, faceEmbeddingModel, imageFactory, defaultImagePipeline, sentenceTokenizer) : null;
   }
 
   @Bean(name = "redisJSONKeyValueAdapter")
